Move Otel.Trace.Exporter namespace to just Otel.Trace (#903)
