+++ examples/Console/TestConsoleExporter.cs
@@ -20,7 +20,6 @@ using System.Threading;
 using System.Threading.Tasks;
 using OpenTelemetry.Resources;
 using OpenTelemetry.Trace;
-using OpenTelemetry.Trace.Export;
 
 namespace Examples.Console
 {
