+++ src/OpenTelemetry.Exporter.Console/ConsoleExporter.cs
@@ -23,7 +23,6 @@ using System.Threading;
 using System.Threading.Tasks;
 using OpenTelemetry.Resources;
 using OpenTelemetry.Trace;
-using OpenTelemetry.Trace.Export;
 
 namespace OpenTelemetry.Exporter.Console
 {
