+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/OtlpExporter.cs
@@ -22,7 +22,7 @@ using System.Threading.Tasks;
 using Grpc.Core;
 
 using OpenTelemetry.Exporter.OpenTelemetryProtocol.Implementation;
-using OpenTelemetry.Trace.Export;
+using OpenTelemetry.Trace;
 
 using OtlpCollector = Opentelemetry.Proto.Collector.Trace.V1;
