+++ src/OpenTelemetry/Trace/ActivityProcessor.cs
@@ -17,7 +17,7 @@ using System.Diagnostics;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace OpenTelemetry.Trace.Export
+namespace OpenTelemetry.Trace
 {
     /// <summary>
     /// Activity processor base class.
