+++ src/OpenTelemetry/Trace/Internal/BroadcastActivityProcessor.cs
@@ -22,7 +22,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using OpenTelemetry.Internal;
 
-namespace OpenTelemetry.Trace.Export.Internal
+namespace OpenTelemetry.Trace.Internal
 {
     internal class BroadcastActivityProcessor : ActivityProcessor, IDisposable
     {
