+++ src/OpenTelemetry/Trace/Internal/NoopActivityProcessor.cs
@@ -18,7 +18,7 @@ using System.Diagnostics;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace OpenTelemetry.Trace.Export.Internal
+namespace OpenTelemetry.Trace.Internal
 {
     internal sealed class NoopActivityProcessor : ActivityProcessor
     {
