+++ src/OpenTelemetry/Trace/SimpleActivityProcessor.cs
@@ -19,7 +19,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using OpenTelemetry.Internal;
 
-namespace OpenTelemetry.Trace.Export
+namespace OpenTelemetry.Trace
 {
     /// <summary>
     /// Implements simple activity processor that exports activities in OnEnd call without batching.
