+++ src/OpenTelemetry/Trace/TracerProviderSdk.cs
@@ -18,7 +18,6 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using OpenTelemetry.Resources;
-using OpenTelemetry.Trace.Export;
 
 namespace OpenTelemetry.Trace
 {
