+++ test/OpenTelemetry.Exporter.Jaeger.Tests/TestActivityProcessor.cs
@@ -18,7 +18,7 @@ using System;
 using System.Diagnostics;
 using System.Threading;
 using System.Threading.Tasks;
-using OpenTelemetry.Trace.Export;
+using OpenTelemetry.Trace;
 
 namespace OpenTelemetry.Exporter.Jaeger.Tests
 {
