+++ test/OpenTelemetry.Instrumentation.SqlClient.Tests/SqlClientTests.cs
@@ -28,7 +28,6 @@ using Moq;
 using OpenTelemetry.Instrumentation.SqlClient.Implementation;
 using OpenTelemetry.Internal.Test;
 using OpenTelemetry.Trace;
-using OpenTelemetry.Trace.Export;
 using Xunit;
 
 namespace OpenTelemetry.Instrumentation.SqlClient.Tests
