+++ test/OpenTelemetry.Tests/Implementation/Testing/Export/TestActivityProcessor.cs
@@ -18,7 +18,7 @@ using System;
 using System.Diagnostics;
 using System.Threading;
 using System.Threading.Tasks;
-using OpenTelemetry.Trace.Export;
+using OpenTelemetry.Trace;
 
 namespace OpenTelemetry.Tests.Implementation.Testing.Export
 {
