+++ test/OpenTelemetry.Tests/Implementation/Trace/Config/ActivityProcessorPipelineTests.cs
@@ -21,8 +21,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using OpenTelemetry.Testing.Export;
 using OpenTelemetry.Trace;
-using OpenTelemetry.Trace.Export;
-using OpenTelemetry.Trace.Export.Internal;
+using OpenTelemetry.Trace.Internal;
 using Xunit;
 
 namespace OpenTelemetry.Tests.Impl.Trace.Config
