+++ test/OpenTelemetry.Tests/Implementation/Trace/Config/BroadcastActivityProcessorTests.cs
@@ -17,8 +17,8 @@
 using System;
 using System.Diagnostics;
 using OpenTelemetry.Tests.Implementation.Testing.Export;
-using OpenTelemetry.Trace.Export;
-using OpenTelemetry.Trace.Export.Internal;
+using OpenTelemetry.Trace;
+using OpenTelemetry.Trace.Internal;
 using Xunit;
 
 namespace OpenTelemetry.Tests.Impl.Trace.Config
