+++ test/OpenTelemetry.Tests/Implementation/Trace/Export/BatchingActivityProcessorTests.cs
@@ -25,7 +25,7 @@ using OpenTelemetry.Tests;
 using OpenTelemetry.Trace.Samplers;
 using Xunit;
 
-namespace OpenTelemetry.Trace.Export.Test
+namespace OpenTelemetry.Trace.Test
 {
     public class BatchingActivityProcessorTests : IDisposable
     {
